Effects: Move all reusable text effects to Rendering/Effects
Modifies:
- TUI.vcxproj
- TUI.vcxproj.filters

Adds:
- Rendering/Effects/DigitalRainEffect.h/.cpp
- Rendering/Effects/Donut3D.h/.cpp
- Rendering/Effects/FireEffect.h/.cpp
- Rendering/Effects/WaterWaveEffect.h/.cpp
- Screens/Content/.gitkeep

Removes:
- Screens/Content/DigitalRainEffect.h/.cpp
- Screens/Content/Donut3D.h/.cpp
- Screens/Content/FireEffect.h/.cpp
- Screens/Content/WaterWaveEffect.h/.cpp
