+++ TUI/TUI.vcxproj
@@ -256,10 +256,10 @@
     <ClInclude Include="Rendering\Styles\Themes.h" />
     <ClInclude Include="Rendering\VtFrameEmitter.h" />
     <ClInclude Include="Rendering\VTRun.h" />
-    <ClInclude Include="Screens\Content\DigitalRainEffect.h" />
-    <ClInclude Include="Screens\Content\Donut3D.h" />
-    <ClInclude Include="Screens\Content\FireEffect.h" />
-    <ClInclude Include="Screens\Content\WaterWaveEffect.h" />
+    <ClInclude Include="Rendering\Effects\DigitalRainEffect.h" />
+    <ClInclude Include="Rendering\Effects\Donut3D.h" />
+    <ClInclude Include="Rendering\Effects\FireEffect.h" />
+    <ClInclude Include="Rendering\Effects\WaterWaveEffect.h" />
     <ClInclude Include="Screens\DigitalRainScreen.h" />
     <ClInclude Include="Screens\Donut3DScreen.h" />
     <ClInclude Include="Screens\FireScreen.h" />
@@ -383,10 +383,10 @@
     </ClCompile>
     <ClCompile Include="Rendering\TerminalSessionOptions.cpp" />
     <ClCompile Include="Rendering\VtFrameEmitter.cpp" />
-    <ClCompile Include="Screens\Content\DigitalRainEffect.cpp" />
-    <ClCompile Include="Screens\Content\Donut3D.cpp" />
-    <ClCompile Include="Screens\Content\FireEffect.cpp" />
-    <ClCompile Include="Screens\Content\WaterWaveEffect.cpp" />
+    <ClCompile Include="Rendering\Effects\DigitalRainEffect.cpp" />
+    <ClCompile Include="Rendering\Effects\Donut3D.cpp" />
+    <ClCompile Include="Rendering\Effects\FireEffect.cpp" />
+    <ClCompile Include="Rendering\Effects\WaterWaveEffect.cpp" />
     <ClCompile Include="Screens\DigitalRainScreen.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">false</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|x64'">false</ExcludedFromBuild>
@@ -428,6 +428,7 @@
     <ClCompile Include="Utilities\Unicode\UnicodeWidth.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="Screens\Content\.gitkeep" />
     <None Include="startup.config" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
